requirements.txt
Flask
streamlit
google-generativeai  
